* feature/ISSUE85-MTDirTraversal: (76 commits)
  sync_queue commenting/cleanup.
  New Feature: Multithreaded directory tree traversal.
  Added test cases for normal directory tree traversal, comparing against grep's results.
  Comment.
  Cleanup.
  Mutex around Logger's writes to cerr.
  OSX interleaves characters streamed to cerr from multiple threads.
  OSX portability issues with $MKDIR_P, switched to AS_MKDIR_P().
  Dump logs.
  Added includes for dev_t, ino_t.
  Completed single- and multi-threaded test cases.
  Handling symlink cycle detection ourselves in the multithreaded case.
  Changed wording of recursive dir loop warning to match grep's output.
  Interim.
  Test updates.
  Added dir_tree_traversal.at.  Cleanup of main.cpp.
  Globber::Run() no longer run in a separate thread, it spawns its worker threads.  Some sync_queue efficiency improvements.
  Comment resurrection.
  Cleanup.
  Added command line param to control number of directory traversal threads.
  ...

+++ src/Match.cpp
@@ -21,7 +21,7 @@
 
 #include <algorithm>
 
-Match::Match(const char *start_of_array, size_t array_size, size_t match_start_offset, size_t match_end_offset, long long line_number)
+Match::Match(const char *start_of_array, size_t array_size, size_t match_start_offset, size_t match_end_offset, size_t line_number)
 {
 	auto line_ending = "\n";
 	// Find the start of the line.
